fix(assistant): self-review remediation — vision reason stamp, gemini abort gate, shared error patterns
